Issue #62: Move color function to color_palette.py
