+++ model_inference/color_palette.py
@@ -14,6 +14,7 @@ SET1 = [
     (0.6,                 0.6,                 0.6)
 ]
 
+
 SET2 = [
     (0.4,                 0.76078431372549016, 0.6470588235294118 ),
     (0.9882352941176471,  0.55294117647058827, 0.3843137254901961 ),
@@ -24,6 +25,19 @@ SET2 = [
     (0.89803921568627454, 0.7686274509803922,  0.58039215686274515),
     (0.70196078431372544, 0.70196078431372544, 0.70196078431372544)
 ]
+
+
+def hex_format(color: tuple):
+    hex = ""
+    for v in color:
+        hex += "{:02X}".format(int(v * 255))
+
+    return f"#{hex}"
+
+
+def rgb_format(color: tuple):
+    r, g, b = color
+    return (r*255, g*255, b*255)
 
 # wes_anderson = {
 #   "BottleRocket1": ("#A42820", "#5F5647", "#9B110E", "#3F5151", "#4E2A1E", "#550307", "#0C1707"),
